+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+##0.4.12
+###Improvement
+
+- Added ValueChange event after blur DateField
+###New feature
+
+- In FDate added static equals method for comparing two Dates 
+
 ##0.4.11
 ###Bug fixes: